Cherry picked relevant changes
Fix group rect to fit expectations
Do not blend preview images
Prevent genetation on NaN values pixels
Fix JXL_ASSERT(exp >= 0)
Subnormal values with same exp_bits as output (8) are not eligible for normalization.
Fix negative shift
Unnecessary casting to size_t fools the min-max check.
Fix repetetive ReadHybridUintClustered nesting
Fix noise decoding to RGB(A)8 and add tests.
Fix touching uninitialized values on tiny modular images
Fix faulty DASSERT.
fix incrementing file_pos when jbrd box
Avoid UB in CopyImageTo when xsize == 0.
Fix remaining quant_weights division by zero fuzzer bugs.
Check for invalid denominator in RAW quant tables.
Check invariant to prevent potential heap-buffer-overflow
Fuzz: fix touching not-yet-decoded values.
Fuzz: add early check for fp + modular + XYB combination

+++ lib/jxl/dec_modular.h
@@ -98,8 +98,8 @@ class ModularFrameDecoder {
   void Init(const FrameDimensions& frame_dim) { this->frame_dim = frame_dim; }
   Status DecodeGlobalInfo(BitReader* reader, const FrameHeader& frame_header,
                           bool allow_truncated_group = false);
-  Status DecodeGroup(const Rect& rect, BitReader* reader, size_t minShift,
-                     size_t maxShift, const ModularStreamId& stream,
+  Status DecodeGroup(const Rect& rect, BitReader* reader, int minShift,
+                     int maxShift, const ModularStreamId& stream,
                      bool zerofill);
   // Decodes a VarDCT DC group (`group_id`) from the given `reader`.
   Status DecodeVarDCTDC(size_t group_id, BitReader* reader,
